+++ babel.config.js
@@ -8,21 +8,10 @@ module.exports = {
     [
       '@babel/env',
       {
-        targets: {
-          node: node.substring(2) // Strip `>=`
-        },
-        modules: process.env.MODULE ? false : 'commonjs',
+        targets: { node: node.substring(2) }, // Strip `>=`
+        modules: process.env.BABEL_ESM ? false : 'commonjs',
         shippedProposals: true,
-        useBuiltIns: 'usage'
-      }
-    ]
-  ],
-  plugins: [
-    [
-      '@babel/transform-runtime',
-      {
-        polyfill: false,
-        regenerator: false
+        loose: true
       }
     ]
   ]